fix(llm): drop .int() on PickSchema.ingredientIndex + surface upstream errors
Two changes verified end-to-end against https://alicante.com.ar/receta/263-receta-empanadas-de-pollo/ on Jumbo, returning 10 matched items in 63s.

1) Zod v4's z.number().int() / z.int() emits safe-integer minimum/maximum (-2^53+1, 2^53-1) in its JSON Schema. Anthropic's structured-output validator rejects ANY min/max on integer types. Switching to plain z.number(); the prompt instructs the model to return whole-number indices and the pipeline's Map<number, Pick> lookup works either way.

2) extract.ts and match.ts previously surfaced only error.message ('Provider returned error'), hiding Anthropic's actual rejection reason. New lib/llm/errors.ts pulls responseBody / data / cause into the LLM_FAILED detail and console.errors the full error object server-side.

+++ src/lib/llm/types.ts
@@ -10,12 +10,14 @@ export type Ingredient = z.infer<typeof IngredientSchema>;
 
 export const ExtractSchema = z.array(IngredientSchema);
 
-// NOTE: do NOT add `.min()`/`.max()`/`.nonnegative()` etc. on numeric fields.
-// Anthropic's structured-output validator rejects JSON Schemas that include
-// `minimum`/`maximum` on `integer` types (error: "For 'integer' type, properties
-// maximum, minimum are not supported"). Validate ranges in code instead.
+// NOTE: use `z.number()`, NOT `z.number().int()`/`z.int()`. Zod v4's int() emits
+// safe-integer `minimum`/`maximum` bounds in its JSON Schema, and Anthropic's
+// structured-output validator rejects min/max on `integer` types ("For 'integer'
+// type, properties maximum, minimum are not supported"). The prompt instructs the
+// model to return whole-number indices; the pipeline's Map<number, Pick> lookup
+// works regardless.
 export const PickSchema = z.object({
-  ingredientIndex: z.number().int(),
+  ingredientIndex: z.number(),
   pickedSkuId: z.string().nullable(),
   confidence: z.enum(['high', 'medium', 'low']),
   reason: z.string(),
